requirements-dev.txt
black==23.10.1
    # via -r requirements-dev.in
boto3==1.28.71
    # via -r requirements.in
boto3-stubs==1.28.71
    # via -r requirements-dev.in
botocore==1.31.71
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.31.71
    # via boto3-stubs
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mypy==1.6.1
    # via -r requirements-dev.in
mypy-boto3-sqs==1.28.36
    # via boto3-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.2
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements-dev.in
platformdirs==3.11.0
    # via black
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0.1
    # via -r requirements.in
s3transfer==0.7.0
    # via boto3
six==1.16.0
    # via python-dateutil
types-awscrt==0.19.3
    # via botocore-stubs
types-pyyaml==6.0.12.12
    # via -r requirements-dev.in
types-s3transfer==0.7.0
    # via boto3-stubs
typing-extensions==4.8.0
    # via
    #   boto3-stubs
    #   mypy
    #   mypy-boto3-sqs
urllib3==2.0.7
    # via botocore
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
